refactor: centralize utilities and optimize timeline
